Simplify URL placeholder substitution code
